% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrq_glasso_coef.R
\name{coef.cv.hrq_glasso}
\alias{coef.cv.hrq_glasso}
\title{Extract coefficients from cv.hrq_glasso object}
\usage{
\method{coef}{cv.hrq_glasso}(object, s, ...)
}
\arguments{
\item{object}{The model object \code{cv.hrq_glasso object}.}

\item{s}{Value of lambda.}

\item{...}{other input parameters.}
}
\value{
The function extract estimated coefficients from \code{cv.hrq_glasso} object.
}
\description{
Extract coefficients from cv.hrq_glasso object
}
