% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrq_glasso_others.R
\name{predict.cv.hrq_glasso}
\alias{predict.cv.hrq_glasso}
\title{Prediction for the cv.hrq_glasso object}
\usage{
\method{predict}{cv.hrq_glasso}(object, newX, s, ...)
}
\arguments{
\item{object}{The model object of \code{hrq_glasso}.}

\item{newX}{New design matrix.}

\item{s}{Value of lambda.}

\item{...}{other input parameters.}
}
\value{
The function returns predicted values based on the fitted model from \code{cv.hrq_glasso}.
}
\description{
Prediction for the cv.hrq_glasso object
}
