% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cores.R
\name{solvebeta}
\alias{solvebeta}
\title{Core function for updating beta}
\usage{
solvebeta(
  x,
  y,
  tau,
  gamma,
  weights,
  group.index,
  lambdaj,
  w.lambda,
  eigenval,
  beta.ini,
  max_iter,
  apprx,
  epsilon = 1e-04
)
}
\arguments{
\item{x}{Design matrix (in matrix format)}

\item{y}{Response variable}

\item{tau}{Percentile}

\item{gamma}{Huber parameter}

\item{weights}{Observation weights}

\item{group.index}{A vector of group index, e.g., (1,1,1,2,2,2,3,3)}

\item{lambdaj}{Shrinkage parameter}

\item{w.lambda}{Weights for Shrinkage parameter of each group}

\item{eigenval}{Largest eigenvalue of the submatrix of H corresponding to each group}

\item{beta.ini}{Initial or previously updated beta}

\item{max_iter}{Maximum number of iteration}

\item{apprx}{Approximation method}

\item{epsilon}{Convergence criterion}
}
\value{
\item{beta.update}{the updated coefficient estimates}
\item{converge}{indicator of convergence}
\item{r}{updated residual}
\item{iter}{number of iteration}
\item{grad}{gradient vector}
}
\description{
Core function for updating beta
}
