% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrq_glasso_predict.R
\name{predict.cv.hrq_glasso}
\alias{predict.cv.hrq_glasso}
\title{Prediction for cv.hrq_glasso object}
\usage{
\method{predict}{cv.hrq_glasso}(object, newX, s, ...)
}
\arguments{
\item{object}{The model object of \code{cv.hrq_glasso}.}

\item{newX}{New design matrix.}

\item{s}{Value of lambda. If missing, the default is the \code{lambda.min}.}

\item{...}{other input parameters.}
}
\value{
The function returns predicted values based on the fitted model from \code{cv.hrq_glasso}.
}
\description{
Prediction for cv.hrq_glasso object
}
